requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
docopt==0.6.2
easywatch==0.0.5
fire==0.4.0
idna==2.10
iniconfig==1.1.1
isodate==0.6.0
Jinja2==2.11.3
jsonschema==3.2.0
libsass==0.20.1
Markdown==3.3.4
MarkupSafe==1.1.1
packaging==20.9
petl==1.7.2
pluggy==0.13.1
py==1.10.0
pynpm==0.1.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
python-slugify==4.0.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
shellingham==1.4.0
simpleeval==0.9.10
six==1.15.0
staticjinja==1.0.4
stringcase==1.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typer==0.3.2
urllib3==1.26.3
validators==0.18.2
watchdog==2.0.2
